requirements.txt
# sanjai-insight Dependencies

# Core
anthropic>=0.39.0
google-generativeai>=0.8.0
python-telegram-bot>=21.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0

# Data & Config
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.20.0

# Scheduling
apscheduler>=3.10.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.1.0

# Load Testing & Automation
locust>=2.15.0
httpx>=0.27.0

# Reporting
jinja2>=3.1.0

# Web Server
fastapi>=0.115.0
uvicorn>=0.32.0

# System Metrics
psutil>=6.1.0
